Move grass if a player is near
I inititally thought about sending some kind of wind from the server
That can be use to have idle animations of grass

But can also be used to move the grass when a player is inside.
This way the server could not send the player at all when hidden in the
grass.

The advantage of that is that a hacked client can not distinguish wind
and player movement in high grass.

But then players would disappear completly even if their foot would be
visible or if the grass density is too low and they would peak through.

So for now lets not overcomplicate the network protocol and bandwith
just to reduce cheating. Because even then a hacked client can highlight
all wind positions. So either wind has to be confusingly similar to
players hiding in grass which would also be annoying for legit players.